Refactor: Modularize code and improve function naming
