+++ cli/bin/config.js
@@ -1,0 +1,4 @@
+import os from 'os';
+
+export const HOME_DIR = os.homedir();
+export const NWC_PATH = `${HOME_DIR}/.pkgzap`;
